--- src/main/java/saphy/saphy/cart/domain/Cart.java
@@ -1,29 +1,0 @@
-package saphy.saphy.cart.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import saphy.saphy.global.entity.BaseEntity;
-import saphy.saphy.item.domain.Item;
-import saphy.saphy.member.domain.Member;
-
-@Entity
-@Getter @Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "cart")
-public class Cart extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id")
-    private Item item;
-}
