--- src/main/java/saphy/saphy/cart/domain/repository/CartRepository.java
@@ -1,8 +1,0 @@
-package saphy.saphy.cart.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import saphy.saphy.cart.domain.Cart;
-
-public interface CartRepository extends JpaRepository<Cart, Long> {
-
-}
