--- src/main/java/saphy/saphy/cart/presentation/CartController.java
@@ -1,11 +1,0 @@
-package saphy.saphy.cart.presentation;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping
-public class CartController {
-}
