+++ src/main/java/saphy/saphy/contact/domain/Category.java
@@ -1,4 +1,4 @@
-package saphy.saphy.productInquiry.domain;
+package saphy.saphy.contact.domain;
 
 public enum Category {
     SMARTPHONE,
