+++ src/main/java/saphy/saphy/contact/domain/Contact.java
@@ -1,4 +1,4 @@
-package saphy.saphy.repairCenterBookmark.domain;
+package saphy.saphy.contact.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -6,23 +6,34 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import saphy.saphy.global.entity.BaseEntity;
-import saphy.saphy.repairCenter.domain.RepairCenter;
 import saphy.saphy.member.domain.Member;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "repair_center_bookmark")
-public class RepairCenterBookmark extends BaseEntity {
+@Table(name = "product_inquiry")
+public class Contact extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "repair_center_id")
-    private RepairCenter repairCenter;
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Category category;
+
+    @Column(nullable = false)
+    private String status;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String question;
+
+    @Column(nullable = false)
+    private String answer;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
