+++ src/main/java/saphy/saphy/coupon/domain/Coupon.java
@@ -1,15 +1,15 @@
 package saphy.saphy.coupon.domain;
 
 import jakarta.persistence.*;
+import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import saphy.saphy.global.entity.BaseEntity;
 import saphy.saphy.member.domain.Member;
 
 @Entity
-@Getter @Setter
+@Getter 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "coupon")
 public class Coupon extends BaseEntity {
@@ -18,18 +18,27 @@ public class Coupon extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private String rate;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private CouponType couponType;
+    @Column
+    private String discountRate; // 할인 퍼센트  ex:30%
 
     @Column
-    private double minimumAmount;
+    private String discountAmount; // 할인 금액 ex:3000원
 
-    @Column(nullable = false)
-    private String content;
+    @Enumerated(EnumType.STRING)
+    @Column
+    private CouponType couponType; // 쿠폰 종류
+
+    @Column
+    private double minimumPrice; // 쿠폰 적용이 가능한 최소 주문 금액
+
+    @Column
+    private String name; // 쿠폰 이름
+    
+    @Column
+    private String content; // 쿠폰 설명
+
+    @Column
+    private LocalDate endAt; // 만료 날짜
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
