--- src/main/java/saphy/saphy/delivery/domain/Delivery.java
@@ -1,40 +1,0 @@
-package saphy.saphy.delivery.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import saphy.saphy.global.entity.BaseEntity;
-import saphy.saphy.order.domain.Order;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter @Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "delivery")
-public class Delivery extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private DeliveryStatus deliveryStatus;
-
-    @Column(nullable = false)
-    private String currentLocation;
-
-    @Column(nullable = false)
-    private LocalDateTime startAt;
-
-    @Column(nullable = false)
-    private LocalDateTime endAt;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id")
-    private Order order;
-
-}
