--- src/main/java/saphy/saphy/delivery/domain/DeliveryStatus.java
@@ -1,7 +1,0 @@
-package saphy.saphy.delivery.domain;
-
-public enum DeliveryStatus {
-    STARTED,    // 배달 시작
-    GOING, // 배달 중
-    DELIVERED   // 배달 완료
-}
