--- src/main/java/saphy/saphy/delivery/presentation/DeliveryController.java
@@ -1,12 +1,0 @@
-package saphy.saphy.delivery.presentation;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping
-public class DeliveryController {
-
-}
