--- src/main/java/saphy/saphy/productInquiry/domain/repository/ProductInquiryRepository.java
@@ -1,8 +1,0 @@
-package saphy.saphy.productInquiry.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import saphy.saphy.productInquiry.domain.ProductInquiry;
-
-public interface ProductInquiryRepository extends JpaRepository<ProductInquiry, Long> {
-
-}
