--- src/main/java/saphy/saphy/productInquiry/service/ProductInquiryService.java
@@ -1,12 +1,0 @@
-package saphy.saphy.productInquiry.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class ProductInquiryService {
-
-}
