+++ src/main/java/saphy/saphy/repairCenter/domain/RepairCenterBookmark.java
@@ -1,4 +1,4 @@
-package saphy.saphy.productInquiry.domain;
+package saphy.saphy.repairCenter.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -6,34 +6,23 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import saphy.saphy.global.entity.BaseEntity;
+import saphy.saphy.repairCenter.domain.RepairCenter;
 import saphy.saphy.member.domain.Member;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "product_inquiry")
-public class ProductInquiry extends BaseEntity {
+@Table(name = "repair_center_bookmark")
+public class RepairCenterBookmark extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Category category;
-
-    @Column(nullable = false)
-    private String status;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String question;
-
-    @Column(nullable = false)
-    private String answer;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "repair_center_id")
+    private RepairCenter repairCenter;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
